training_code/requirements.txt
opencv-python
pyyaml
tqdm
decord
webdataset
triton
lpips
av
# flash-attn
safetensors 